capabilitiesV2: refactor product reconciler
