+++ pkg/controller/product/3scale_reconciler.go
@@ -1,9 +1,7 @@
 package product
 
 import (
-	"errors"
 	"fmt"
-	"strconv"
 
 	capabilitiesv1beta1 "github.com/3scale/3scale-operator/pkg/apis/capabilities/v1beta1"
 	"github.com/3scale/3scale-operator/pkg/helper"
@@ -12,31 +10,6 @@ import (
 	threescaleapi "github.com/3scale/3scale-porta-go-client/client"
 	"github.com/go-logr/logr"
 )
-
-type methodData struct {
-	item threescaleapi.MethodItem
-	spec capabilitiesv1beta1.Methodpec
-}
-
-type metricData struct {
-	item threescaleapi.MetricItem
-	spec capabilitiesv1beta1.MetricSpec
-}
-
-type mappingRuleData struct {
-	item threescaleapi.MappingRuleItem
-	spec capabilitiesv1beta1.MappingRuleSpec
-}
-
-type backendUsageData struct {
-	item threescaleapi.BackendAPIUsageItem
-	spec capabilitiesv1beta1.BackendUsageSpec
-}
-
-type newBackendUsageData struct {
-	item threescaleapi.BackendApiItem
-	spec capabilitiesv1beta1.BackendUsageSpec
-}
 
 type ThreescaleReconciler struct {
 	*reconcilers.BaseReconciler
@@ -150,479 +123,6 @@ func (t *ThreescaleReconciler) syncProduct(_ interface{}) error {
 	return nil
 }
 
-func (t *ThreescaleReconciler) syncMethods(_ interface{}) error {
-	desiredKeys := make([]string, 0, len(t.resource.Spec.Methods))
-	for systemName := range t.resource.Spec.Methods {
-		desiredKeys = append(desiredKeys, systemName)
-	}
-
-	existingMap := map[string]threescaleapi.MethodItem{}
-	existingList, err := t.entity.Methods()
-	if err != nil {
-		return fmt.Errorf("Error sync product methods [%s]: %w", t.resource.Spec.SystemName, err)
-	}
-
-	existingKeys := make([]string, 0, len(existingList.Methods))
-	for _, existing := range existingList.Methods {
-		systemName := existing.Element.SystemName
-		existingKeys = append(existingKeys, systemName)
-		existingMap[systemName] = existing.Element
-	}
-
-	desiredNewKeys := helper.ArrayStringDifference(desiredKeys, existingKeys)
-	desiredNewMap := map[string]capabilitiesv1beta1.Methodpec{}
-	for _, systemName := range desiredNewKeys {
-		// key is expected to exist
-		// desiredNewKeys is a subset of the Spec.Method map key set
-		desiredNewMap[systemName] = t.resource.Spec.Methods[systemName]
-	}
-	err = t.createNewMethods(desiredNewMap)
-	if err != nil {
-		return fmt.Errorf("Error sync product methods [%s]: %w", t.resource.Spec.SystemName, err)
-	}
-
-	notDesiredExistingKeys := helper.ArrayStringDifference(existingKeys, desiredKeys)
-	notDesiredMap := map[string]threescaleapi.MethodItem{}
-	for _, systemName := range notDesiredExistingKeys {
-		// key is expected to exist
-		// notDesiredExistingKeys is a subset of the existingMap key set
-		notDesiredMap[systemName] = existingMap[systemName]
-	}
-	err = t.processNotDesiredMethods(notDesiredMap)
-	if err != nil {
-		return fmt.Errorf("Error sync product methods [%s]: %w", t.resource.Spec.SystemName, err)
-	}
-
-	matchedKeys := helper.ArrayStringIntersection(existingKeys, desiredKeys)
-	matchedMap := map[string]methodData{}
-	for _, systemName := range matchedKeys {
-		matchedMap[systemName] = methodData{
-			item: existingMap[systemName],
-			spec: t.resource.Spec.Methods[systemName],
-		}
-	}
-
-	err = t.reconcileMatchedMethods(matchedMap)
-	if err != nil {
-		return fmt.Errorf("Error sync product methods [%s]: %w", t.resource.Spec.SystemName, err)
-	}
-
-	return nil
-}
-
-func (t *ThreescaleReconciler) createNewMethods(desiredNewMap map[string]capabilitiesv1beta1.Methodpec) error {
-	for systemName, method := range desiredNewMap {
-		params := threescaleapi.Params{
-			"friendly_name": method.Name,
-			"system_name":   systemName,
-		}
-		if len(method.Description) > 0 {
-			params["description"] = method.Description
-		}
-		err := t.entity.CreateMethod(params)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (t *ThreescaleReconciler) processNotDesiredMethods(notDesiredMap map[string]threescaleapi.MethodItem) error {
-	for _, notDesiredMethod := range notDesiredMap {
-		err := t.entity.DeleteMethod(notDesiredMethod.ID)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (t *ThreescaleReconciler) reconcileMatchedMethods(matchedMap map[string]methodData) error {
-	for _, data := range matchedMap {
-		params := threescaleapi.Params{}
-		if data.spec.Name != data.item.Name {
-			params["friendly_name"] = data.spec.Name
-		}
-
-		if data.spec.Description != data.item.Description {
-			params["description"] = data.spec.Description
-		}
-
-		if len(params) > 0 {
-			err := t.entity.UpdateMethod(data.item.ID, params)
-			if err != nil {
-				return fmt.Errorf("Error reconcile product methods: %w", err)
-			}
-		}
-	}
-
-	return nil
-}
-
-func (t *ThreescaleReconciler) syncMetrics(_ interface{}) error {
-	desiredKeys := make([]string, 0, len(t.resource.Spec.Metrics))
-	for systemName := range t.resource.Spec.Metrics {
-		desiredKeys = append(desiredKeys, systemName)
-	}
-
-	existingMap := map[string]threescaleapi.MetricItem{}
-	existingList, err := t.entity.Metrics()
-	if err != nil {
-		return fmt.Errorf("Error sync product metrics [%s]: %w", t.resource.Spec.SystemName, err)
-	}
-
-	existingKeys := make([]string, 0, len(existingList.Metrics))
-	for _, existing := range existingList.Metrics {
-		systemName := existing.Element.SystemName
-		existingKeys = append(existingKeys, systemName)
-		existingMap[systemName] = existing.Element
-	}
-
-	//
-	// Deleted existing and not desired metrics
-	//
-
-	notDesiredExistingKeys := helper.ArrayStringDifference(existingKeys, desiredKeys)
-	notDesiredMap := map[string]threescaleapi.MetricItem{}
-	for _, systemName := range notDesiredExistingKeys {
-		// key is expected to exist
-		// notDesiredExistingKeys is a subset of the existingMap key set
-		notDesiredMap[systemName] = existingMap[systemName]
-	}
-	err = t.processNotDesiredMetrics(notDesiredMap)
-	if err != nil {
-		return fmt.Errorf("Error sync product metrics [%s]: %w", t.resource.Spec.SystemName, err)
-	}
-
-	//
-	// Reconcile existing and changed metrics
-	//
-
-	matchedKeys := helper.ArrayStringIntersection(existingKeys, desiredKeys)
-	matchedMap := map[string]metricData{}
-	for _, systemName := range matchedKeys {
-		matchedMap[systemName] = metricData{
-			item: existingMap[systemName],
-			spec: t.resource.Spec.Metrics[systemName],
-		}
-	}
-
-	err = t.reconcileMatchedMetrics(matchedMap)
-	if err != nil {
-		return fmt.Errorf("Error sync product metrics [%s]: %w", t.resource.Spec.SystemName, err)
-	}
-
-	//
-	// Create not existing and desired metrics
-	//
-
-	desiredNewKeys := helper.ArrayStringDifference(desiredKeys, existingKeys)
-	desiredNewMap := map[string]capabilitiesv1beta1.MetricSpec{}
-	for _, systemName := range desiredNewKeys {
-		// key is expected to exist
-		// desiredNewKeys is a subset of the Spec.Metrics map key set
-		desiredNewMap[systemName] = t.resource.Spec.Metrics[systemName]
-	}
-	err = t.createNewMetrics(desiredNewMap)
-	if err != nil {
-		return fmt.Errorf("Error sync product metrics [%s]: %w", t.resource.Spec.SystemName, err)
-	}
-
-	return nil
-}
-
-func (t *ThreescaleReconciler) createNewMetrics(desiredNewMap map[string]capabilitiesv1beta1.MetricSpec) error {
-	for systemName, metric := range desiredNewMap {
-		params := threescaleapi.Params{
-			"friendly_name": metric.Name,
-			"unit":          metric.Unit,
-			"system_name":   systemName,
-		}
-		if len(metric.Description) > 0 {
-			params["description"] = metric.Description
-		}
-		err := t.entity.CreateMetric(params)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (t *ThreescaleReconciler) processNotDesiredMetrics(notDesiredMap map[string]threescaleapi.MetricItem) error {
-	for _, metric := range notDesiredMap {
-		err := t.entity.DeleteMetric(metric.ID)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (t *ThreescaleReconciler) reconcileMatchedMetrics(matchedMap map[string]metricData) error {
-	for _, data := range matchedMap {
-		params := threescaleapi.Params{}
-		if data.spec.Name != data.item.Name {
-			params["friendly_name"] = data.spec.Name
-		}
-
-		if data.spec.Unit != data.item.Unit {
-			params["unit"] = data.spec.Unit
-		}
-
-		if data.spec.Description != data.item.Description {
-			params["description"] = data.spec.Description
-		}
-
-		if len(params) > 0 {
-			err := t.entity.UpdateMetric(data.item.ID, params)
-			if err != nil {
-				return fmt.Errorf("Error updating product metric: %w", err)
-			}
-		}
-	}
-
-	return nil
-}
-
-func (t *ThreescaleReconciler) syncMappingRules(_ interface{}) error {
-	desiredKeys := make([]string, 0, len(t.resource.Spec.MappingRules))
-	desiredMap := map[string]capabilitiesv1beta1.MappingRuleSpec{}
-	for _, spec := range t.resource.Spec.MappingRules {
-		key := fmt.Sprintf("%s:%s", spec.HTTPMethod, spec.Pattern)
-		desiredKeys = append(desiredKeys, key)
-		desiredMap[key] = spec
-	}
-
-	existingKeys := []string{}
-	existingMap := map[string]threescaleapi.MappingRuleItem{}
-	existingList, err := t.entity.MappingRules()
-	if err != nil {
-		return fmt.Errorf("Error sync product [%s] mappingrules: %w", t.resource.Spec.SystemName, err)
-	}
-	for _, item := range existingList.MappingRules {
-		key := fmt.Sprintf("%s:%s", item.Element.HTTPMethod, item.Element.Pattern)
-		existingKeys = append(existingKeys, key)
-		existingMap[key] = item.Element
-	}
-
-	//
-	// Deleted existing and not desired mapping rules
-	//
-
-	notDesiredExistingKeys := helper.ArrayStringDifference(existingKeys, desiredKeys)
-	notDesiredList := make([]threescaleapi.MappingRuleItem, 0, len(notDesiredExistingKeys))
-	for _, key := range notDesiredExistingKeys {
-		// key is expected to exist
-		// notDesiredExistingKeys is a subset of the existingKeys set
-		notDesiredList = append(notDesiredList, existingMap[key])
-	}
-	err = t.processNotDesiredMappingRules(notDesiredList)
-	if err != nil {
-		return fmt.Errorf("Error sync product [%s] mappingrules: %w", t.resource.Spec.SystemName, err)
-	}
-
-	//
-	// Reconcile existing and changed mapping rules
-	//
-
-	matchedKeys := helper.ArrayStringIntersection(existingKeys, desiredKeys)
-	matchedList := make([]mappingRuleData, 0, len(matchedKeys))
-	for _, key := range matchedKeys {
-		matchedList = append(matchedList, mappingRuleData{
-			item: existingMap[key],
-			spec: desiredMap[key],
-		})
-	}
-
-	err = t.reconcileMatchedMappingRules(matchedList)
-	if err != nil {
-		return fmt.Errorf("Error sync product [%s] mappingrules: %w", t.resource.Spec.SystemName, err)
-	}
-
-	//
-	// Create not existing and desired mapping rules
-	//
-
-	desiredNewKeys := helper.ArrayStringDifference(desiredKeys, existingKeys)
-	desiredNewList := make([]capabilitiesv1beta1.MappingRuleSpec, 0, len(desiredNewKeys))
-	for _, key := range desiredNewKeys {
-		// key is expected to exist
-		// desiredNewKeys is a subset of the desiredKeys set
-		desiredNewList = append(desiredNewList, desiredMap[key])
-	}
-	err = t.createNewMappingRules(desiredNewList)
-	if err != nil {
-		return fmt.Errorf("Error sync product [%s] mappingrules: %w", t.resource.Spec.SystemName, err)
-	}
-
-	return nil
-}
-
-func (t *ThreescaleReconciler) processNotDesiredMappingRules(notDesiredList []threescaleapi.MappingRuleItem) error {
-	for _, mappingRule := range notDesiredList {
-		err := t.entity.DeleteMappingRule(mappingRule.ID)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (t *ThreescaleReconciler) reconcileMatchedMappingRules(matchedList []mappingRuleData) error {
-	for _, data := range matchedList {
-		params := threescaleapi.Params{}
-
-		//
-		// Reconcile metric or method
-		//
-		metricID, err := t.entity.FindMethodMetricIDBySystemName(data.spec.MetricMethodRef)
-		if err != nil {
-			return fmt.Errorf("Error reconcile product mapping rule: %w", err)
-		}
-
-		if metricID < 0 {
-			// Should not happen as metric and method references have been validated and should exists
-			return errors.New("product metric method ref for mapping rule not found")
-		}
-
-		if metricID != data.item.MetricID {
-			params["metric_id"] = strconv.FormatInt(metricID, 10)
-		}
-
-		//
-		// Reconcile delta
-		//
-		if *data.spec.Increment != data.item.Delta {
-			params["delta"] = strconv.Itoa(*data.spec.Increment)
-		}
-
-		if len(params) > 0 {
-			err := t.entity.UpdateMappingRule(data.item.ID, params)
-			if err != nil {
-				return fmt.Errorf("Error reconcile product mapping rule: %w", err)
-			}
-		}
-	}
-
-	return nil
-}
-
-func (t *ThreescaleReconciler) createNewMappingRules(desiredList []capabilitiesv1beta1.MappingRuleSpec) error {
-	for _, spec := range desiredList {
-		metricID, err := t.entity.FindMethodMetricIDBySystemName(spec.MetricMethodRef)
-		if err != nil {
-			return fmt.Errorf("Error creating product [%s] mappingrule: %w", t.resource.Spec.SystemName, err)
-		}
-
-		if metricID < 0 {
-			// Should not happen as metric and method references have been validated and should exists
-			return errors.New("product metric method ref for mapping rule not found")
-		}
-
-		params := threescaleapi.Params{
-			"pattern":     spec.Pattern,
-			"http_method": spec.HTTPMethod,
-			"metric_id":   strconv.FormatInt(metricID, 10),
-			// Defaults are set in the spec, should not be nil
-			"delta": strconv.Itoa(*spec.Increment),
-		}
-
-		err = t.entity.CreateMappingRule(params)
-		if err != nil {
-			return fmt.Errorf("Error creating product [%s] mappingrule: %w", t.resource.Spec.SystemName, err)
-		}
-	}
-	return nil
-}
-
-func (t *ThreescaleReconciler) syncBackendUsage(_ interface{}) error {
-	backendAPIs, err := t.threescaleAPIClient.ListBackendApis()
-	if err != nil {
-		return fmt.Errorf("Error sync product [%s] backendusages: %w", t.resource.Spec.SystemName, err)
-	}
-
-	backendMapByID := map[int64]threescaleapi.BackendApiItem{}
-	backendMapBySystemName := map[string]threescaleapi.BackendApiItem{}
-	for _, backend := range backendAPIs.Backends {
-		backendMapByID[backend.Element.ID] = backend.Element
-		backendMapBySystemName[backend.Element.SystemName] = backend.Element
-	}
-
-	desiredKeys := make([]string, 0, len(t.resource.Spec.BackendUsages))
-	for systemName := range t.resource.Spec.BackendUsages {
-		desiredKeys = append(desiredKeys, systemName)
-	}
-
-	existingList, err := t.entity.BackendUsages()
-	if err != nil {
-		return fmt.Errorf("Error sync product [%s] backendusages: %w", t.resource.Spec.SystemName, err)
-	}
-
-	existingKeys := make([]string, 0, len(existingList))
-	existingMap := map[string]threescaleapi.BackendAPIUsageItem{}
-	for _, existing := range existingList {
-		// backend usage ID should exist in the backend list
-		systemName := backendMapByID[existing.Element.BackendAPIID].SystemName
-		existingKeys = append(existingKeys, systemName)
-		existingMap[systemName] = existing.Element
-	}
-
-	//
-	// Deleted existing and not desired
-	//
-	notDesiredExistingKeys := helper.ArrayStringDifference(existingKeys, desiredKeys)
-	notDesiredList := make([]threescaleapi.BackendAPIUsageItem, 0, len(notDesiredExistingKeys))
-	for _, systemName := range notDesiredExistingKeys {
-		// key is expected to exist
-		// notDesiredExistingKeys is a subset of the existingMap key set
-		notDesiredList = append(notDesiredList, existingMap[systemName])
-	}
-	err = t.processNotDesiredBackendUsages(notDesiredList)
-	if err != nil {
-		return fmt.Errorf("Error sync product [%s] backendusages: %w", t.resource.Spec.SystemName, err)
-	}
-
-	//
-	// Reconcile existing and changed
-	//
-	matchedKeys := helper.ArrayStringIntersection(existingKeys, desiredKeys)
-	matchedMap := map[string]backendUsageData{}
-	for _, systemName := range matchedKeys {
-		matchedMap[systemName] = backendUsageData{
-			item: existingMap[systemName],
-			spec: t.resource.Spec.BackendUsages[systemName],
-		}
-	}
-	err = t.reconcileMatchedBackendUsages(matchedMap)
-	if err != nil {
-		return fmt.Errorf("Error sync product [%s] backendusages: %w", t.resource.Spec.SystemName, err)
-	}
-
-	//
-	// Create not existing and desired
-	// Spec validation makes sure all backend resources referenced by backend usage
-	// exist and are sync'ed. Thus, they should exist in the backend entity map
-	desiredNewKeys := helper.ArrayStringDifference(desiredKeys, existingKeys)
-	desiredNewList := make([]newBackendUsageData, 0, len(desiredNewKeys))
-	for _, backendSystemName := range desiredNewKeys {
-		// backendSystemName is expected to exist
-		// desiredNewKeys is a subset of the Spec.BackendUsages map key set
-		desiredNewList = append(desiredNewList, newBackendUsageData{
-			spec: t.resource.Spec.BackendUsages[backendSystemName],
-			// desiredNewKeys is a set of backend resources systemName's
-			// which have been validated of being sync'ed.
-			// Thus, the map of existing Backend entities should contain desired systemName
-			item: backendMapBySystemName[backendSystemName],
-		})
-	}
-	err = t.createNewBackendUsage(desiredNewList)
-	if err != nil {
-		return fmt.Errorf("Error sync product [%s] backendusages: %w", t.resource.Spec.SystemName, err)
-	}
-	return nil
-}
-
 func (t *ThreescaleReconciler) syncProxy(_ interface{}) error {
 	return nil
 }
@@ -636,48 +136,5 @@ func (t *ThreescaleReconciler) syncPolicies(_ interface{}) error {
 }
 
 func (t *ThreescaleReconciler) bumpProxyVersion(_ interface{}) error {
-	return nil
-}
-
-func (t *ThreescaleReconciler) processNotDesiredBackendUsages(notDesiredList []threescaleapi.BackendAPIUsageItem) error {
-	for _, item := range notDesiredList {
-		err := t.entity.DeleteBackendUsage(item.ID)
-		if err != nil {
-			return fmt.Errorf("Error deleting product backendusage: %w", err)
-		}
-	}
-	return nil
-}
-
-func (t *ThreescaleReconciler) reconcileMatchedBackendUsages(matchedMap map[string]backendUsageData) error {
-	for _, data := range matchedMap {
-		params := threescaleapi.Params{}
-		if data.spec.Path != data.item.Path {
-			params["path"] = data.spec.Path
-		}
-
-		if len(params) > 0 {
-			err := t.entity.UpdateBackendUsage(data.item.ID, params)
-			if err != nil {
-				return fmt.Errorf("Error updating product backendusage: %w", err)
-			}
-		}
-	}
-
-	return nil
-}
-
-func (t *ThreescaleReconciler) createNewBackendUsage(matchedList []newBackendUsageData) error {
-	for _, data := range matchedList {
-		params := threescaleapi.Params{
-			"path":           data.spec.Path,
-			"backend_api_id": strconv.FormatInt(data.item.ID, 10),
-		}
-		err := t.entity.CreateBackendUsage(params)
-		if err != nil {
-			return fmt.Errorf("Error creating product backendusage: %w", err)
-		}
-	}
-
 	return nil
 }
